chore: move current entity to sub folder
