+++ packages/backend/test/modules/genres/genres.repository.spec.ts
@@ -1,6 +1,6 @@
 import { DataSource, In, Repository } from 'typeorm';
 
-import { Artwork } from '@/src/modules/artworks/artworks.entity';
+import { Artwork } from '@/src/modules/artworks/entities/artworks.entity';
 import { GenreTranslation } from '@/src/modules/genres/entities/genre-translations.entity';
 import { Genre } from '@/src/modules/genres/entities/genres.entity';
 import { Language } from '@/src/modules/genres/enums/language.enum';
